+++ dataset/example/app/components/examples/treemap/treemap.js
@@ -6,21 +6,15 @@ import webtreemap from 'webtreemap';
 import 'webtreemap/webtreemap.css!';
 import './treemap.css!';
 
-controller.$inject = ['$log'];
-function controller ($log) {
+controller.$inject = ['$log', '$timeout', '$element'];
+function controller ($log, $timeout, $element) {
   const $ctrl = this;
 
-  const $map = angular.element(document.getElementById('_examples_treemap__chart'));
-
   if (screenfull.enabled) {
-    $ctrl.fullscreen = function () {
-      screenfull.request($map[0]);
-    };
-
-    document.addEventListener(screenfull.raw.fullscreenchange, () => {
+    document.addEventListener(screenfull.raw.fullscreenchange, event => {
       $log.debug('screenfull.raw.fullscreenerror');
-      angular.element($map)[screenfull.isFullscreen ? 'addClass' : 'removeClass']('fullscreen');
-      change();
+      angular.element(event.target).toggleClass('fullscreen', screenfull.isFullscreen);
+      draw();
     });
   }
 
@@ -29,80 +23,86 @@ function controller ($log) {
       data: $ctrl.dataPackage,
       enableSvgDownload: false,
       enablePngDownload: false,
-      onChange: change
+      onChange: draw
     },
-    draw: change,
-    $onInit: change
+    draw,
+    $onInit () {
+      $timeout(draw, 0);
+    },
+    fullscreen ($event) {
+      const elm = angular.element($event.target).parent().find('treemap');
+      screenfull.request(elm[0]);
+    }
   });
 
-  function change () {
-    $map.empty();
+  function draw () {
+    angular.forEach($element.find('treemap'), d => {
+      const elm = angular.element(d);
+      const name = elm.attr('data-name');
+      drawMap(elm, $ctrl.dataPackage.resourcesByName[name]);
+    });
+  }
 
-    $ctrl.dataPackage.resources.forEach(res => {
-      const tree = res.data;
-      const treeData = newNode('/');
+  function drawMap (elm, res) {
+    elm.empty();
 
-      if (res.table) {
-        tree.forEach(d => {
-          addNode(d.Source, Number(d.Size), d.Tag);
-        });
-      } else {
-        for (const source in tree) {
-          if (Object.hasOwnProperty.call(tree, source)) {
-            addNode(source, tree[source]);
-          }
+    const tree = res.data;
+    const treeData = newNode('/');
+
+    if (res.table) {
+      tree.forEach(d => {
+        addNode(d.Source, Number(d.Size), d.Tag);
+      });
+    } else {
+      for (const source in tree) {
+        if (Object.hasOwnProperty.call(tree, source)) {
+          addNode(source, tree[source]);
         }
       }
+    }
 
-      addSizeToTitle(treeData, treeData.data.$area);
+    addSizeToTitle(treeData, treeData.data.$area);
+    webtreemap(elm[0], treeData);
+    return;
 
-      function addNode (path, size, tag) {
-        const parts = path.split('/');
-        let node = treeData;
+    function addNode (path, size, tag) {
+      const parts = path.split('/');
+      let node = treeData;
+      node.data.$area += size;
+
+      parts.forEach(part => {
+        let child = node.children.find(child => child.name === part);
+        if (!child) {
+          child = newNode(part, tag);
+          node.children.push(child);
+        }
+
+        node = child;
         node.data.$area += size;
+      });
+    }
 
-        parts.forEach(part => {
-          let child = node.children.find(child => child.name === part);
-          if (!child) {
-            child = newNode(part, tag);
-            node.children.push(child);
-          }
+    function newNode (name, tag) {
+      // var $symbol = (name.slice(-1) === '*') ? 'tag' : '';
+      return {
+        name,
+        data: {
+          $area: 0,
+          symbol: tag
+        },
+        children: []
+      };
+    }
 
-          node = child;
-          node.data.$area += size;
-        });
-      }
+    function addSizeToTitle (node, total) {
+      const size = node.data.$area;
+      const pct = 100.0 * size / total;
 
-      const h = angular.element(`<h4>${res.name}</h4>`);
-      angular.element($map).append(h);
-
-      const d = angular.element('<div></div>');
-      $map.append(d);
-
-      webtreemap(d[0], treeData);
-    });
-  }
-
-  function newNode (name, tag) {
-    // var $symbol = (name.slice(-1) === '*') ? 'tag' : '';
-    return {
-      name,
-      data: {
-        $area: 0,
-        symbol: tag
-      },
-      children: []
-    };
-  }
-
-  function addSizeToTitle (node, total) {
-    const size = node.data.$area;
-    const pct = 100.0 * size / total;
-
-    node.name += ` • ${size.toLocaleString()} • ${pct.toFixed(2)}%`;
-    node.children.forEach(x => {
-      addSizeToTitle(x, total);
-    });
+      node.name += ` • ${size.toLocaleString()} • ${pct.toFixed(2)}%`;
+      node.children.forEach(x => {
+        addSizeToTitle(x, total);
+      });
+    }
   }
 }